// src/types/StudentData.ts
export default interface StudentData {
  user_id: string;
  department_id: string;
  student_status: string;
  student_class: string;
  student_degree: string;
  student_program: string;
  student_major: string;
  email: string;
  user_name: string;
  address: string;
  first_name: string;
  last_name: string;
  birthday: string;
  sex: string;
  description: string;
  country: string;
  department_name: string;
}